--- templates/barriers/history/partials/economic_assessment/assessment.html
@@ -1,6 +1,0 @@
-<div class="history-item__old-value">
-    <span class="diff__del">{{ item.old_value.name }}</span>
-</div>
-<div class="history-item__new-value">
-    <span class="diff__ins">{{ item.new_value.name }}</span>
-</div>
