+++ templates/barriers/history/partials/economic_assessment/rating.html
@@ -1,0 +1,6 @@
+<div class="history-item__old-value">
+    <span class="diff__del">{{ item.old_value }}</span>
+</div>
+<div class="history-item__new-value">
+    <span class="diff__ins">{{ item.new_value }}</span>
+</div>
