+++ utils/metadata.py
@@ -235,16 +235,6 @@ class Metadata:
             if priority["code"] == priority_code:
                 return priority
 
-    def get_assessment_name(self, assessment_code):
-        assessment_names = {
-            "impact": "Economic assessment",
-            "value_to_economy": "Value to UK Economy",
-            "import_market_size": "Import Market Size",
-            "export_value": "Value of currently affected UK exports",
-            "commercial_value": "Commercial Value",
-        }
-        return assessment_names.get(assessment_code)
-
     def get_category_list(self, sort=True):
         """
         Dedupe and sort the barrier types
